CRM-21812 - don't specify installation type post install
props @MegaphoneJon
The wp-cli-login-server plugin checks for the presence of elements in the $_GET array, and fails if any are found. This conflicts with CiviCRM, which places:

$_GET['civicrm_install_type'] = 'wordpress';

into every page load.

I confirmed that this value is only used for install, and placed a conditional around it to only insert pre-install. I successfully installed CiviCRM and engaged in normal use with the conditional applied, and confirmed that this fixed the WP plugin.

This is intended to close aaemnnosttv/wp-cli-login-command#24.

Signed-off-by: Kevin Cristiano <kcristiano@tadpole.cc>