Do not call session_start() on CLI processes
This is a refinement of the idea in https://github.com/civicrm/civicrm-wordpress/pull/135 to address other CLI use-cases.

Before
-------

If I have a `wpmaster` site on on `bknix-max` (php73), and if I run `cv upgrade:db -vv`, then there are several warnings like this:

After
-----

It doesn't try to start session in CLI (where sessions don't make sense).

Signed-off-by: Kevin Cristiano <kcristiano@kcristiano.com>